refactor: 캐릭터 관련 로직 entities 내에서 관리
