+++ entities/character/model/hooks/useGreeting.ts
@@ -1,9 +1,9 @@
+import { fetchGreeting } from "@/entities/character/api/greeting";
 import { Character } from "@/entities/character/model/character";
 import {
 	type CharacterName,
 	DEFAULT_CHARACTER,
 } from "@/entities/character/model/characterMessages";
-import { getGreeting } from "@/features/friends/api/getGreeting";
 import { useAuthStore } from "@/shared/store/auth";
 import { useQuery } from "@tanstack/react-query";
 
@@ -11,7 +11,7 @@ export function useGreeting(context: string = "friend-invite") {
 	const { isLoggedIn } = useAuthStore();
 	const { data: greetingData } = useQuery({
 		queryKey: ["greeting", context],
-		queryFn: () => getGreeting(context),
+		queryFn: () => fetchGreeting({ context }),
 		enabled: !!isLoggedIn,
 	});
