+++ entities/character/store/myCharacter.ts
