+++ features/friends/model/hooks/index.ts
@@ -1,7 +1,6 @@
-export { useFriendRequest } from "./useFriendRequest";
-export { useFriendRequestResponse } from "./useFriendRequestResponse";
-export { useGetFriends } from "./useGetFriends";
-export { useGreeting } from "./useGreeting";
-export { useInviteCodeCopy } from "./useInviteCodeCopy";
-export { useRemoveFriend } from "./useRemoveFriend";
-export { useToast } from "./useToast";
+export { useFriendRequest } from "@/features/friends/model/hooks/useFriendRequest";
+export { useFriendRequestResponse } from "@/features/friends/model/hooks/useFriendRequestResponse";
+export { useGetFriends } from "@/features/friends/model/hooks/useGetFriends";
+export { useInviteCodeCopy } from "@/features/friends/model/hooks/useInviteCodeCopy";
+export { useRemoveFriend } from "@/features/friends/model/hooks/useRemoveFriend";
+export { useToast } from "@/features/friends/model/hooks/useToast";
