+++ features/friends/model/hooks/useGetFriends.ts
@@ -1,7 +1,7 @@
+import { fetchGreeting } from "@/entities/character/api/greeting";
 import { getFriendFeed } from "@/features/friends/api/getFriendFeed";
 import { getFriendRequests } from "@/features/friends/api/getFriendRequests";
 import { getFriends } from "@/features/friends/api/getFriends";
-import { getGreeting } from "@/features/friends/api/getGreeting";
 import { useAuthStore } from "@/shared/store/auth";
 import { useQuery } from "@tanstack/react-query";
 
@@ -28,7 +28,7 @@ export function useGetFriends() {
 
 	const greeting = useQuery({
 		queryKey: ["friends", "greeting", "friend-reminder"],
-		queryFn: () => getGreeting("friend-reminder"),
+		queryFn: () => fetchGreeting({ context: "friend-reminder" }),
 		enabled: !!isLoggedIn,
 	});
